LOL get rid of 90% of HFM
This thing is so badly written I don't even know where to start degrading it.  Maybe
the blatent disregard for safetly closing file streams and causing memory leaks? Maybe
the way nothing is asynchronous to the UI thread and it completely locks your phone up
for over a minute (LOL).  Fuck this thing.  I found a github with a host file updated
weekly so We'll just update the hosts file daily from source.  Keeping the on/off toggle
code though.

Change-Id: I171b1107d50741efb9e710cbd6dcea5cfbb76b7a